Implement filters and avoid requirement objects
The fetcher now takes the package name directly instead of using a
requirement object. We don't want the markers here, and it gets
confusing if we accepts them (in the requirement). Accepting only the
name (as a string) avoids doubts.

Filters are implemented to provide the specifier feature provided by the
requirement. This also gives us more freedom on what can be used to
filter entries.